Moved HookProc to base class Add HookProc to Floating window
